vhost-pci-slave: add "reset_virtio"
After the vhost-pci-net device being hotplugged to the VM, the device
uses the features bits that have been negotiated with the remote virtio
device to negotiate with the driver. If the driver accepts a subset of
the feature bits, it implies that the vhost-pci-net can only suppoort
a subset of the features supported by the remote virtio device. In this
case, the remote virtio_device will be reset, and re-start the vhost-user
protocol.

Add the "reset_virtio" field as an indicator to the slave in this case.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>